+++ commands/__init__.py
@@ -4,5 +4,9 @@ from .help import HelpCommand
 from .start import StartCommand
 from .schedule import ScheduleCommand
 from .watering import WateringCommand
+from .question import QuestionCommand
+from .alcohol70 import Alcohol70Command
+from .drink import DrinkCommand
 
-__all__ = ('StatusCommand', 'StartCommand', 'HelpCommand', 'ScheduleCommand')
+__all__ = ('StatusCommand', 'StartCommand', 'HelpCommand', 'DrinkCommand',
+            'ScheduleCommand', 'QuestionCommand', 'Alcohol70Command')